% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierbase.R
\docType{package}
\name{hierbase}
\alias{hierbase}
\title{hierbase: Enabling Hierarchical Inference}
\description{
The hierbase package provides the functions to perform hierarchical inference.
The main workflow consists of two function calls.
}
\section{Functions}{

The building of the hierarchical tree can be achieved by either of the functions
\code{\link{cluster_vars}} or \code{\link{cluster_positions}}.
The function \code{\link{advance_hierarchy}} (with 'build-in' group test functions) 
and \code{\link{run_hierarchy}} (with a user specified group test function) perform 
the hierarchical testing by going top down through the hierarchical tree and obviously 
require the hierarchical tree as an input.
}

\references{
Renaux, C., Bühlmann, P. (2021), Efficient Multiple Testing 
Adjustment for Hierarchical Inference. <arXiv:2104.15028>
}
